node/redis
Simple node.js server that shows how to use npm as part of the build process (it also connects to the redis server through the created service for post requests).
The pod shows how to expose a service internally and externally at the same time. (not sure if that is the right way to do it, I tried exposing two services with the same selector and that didn't work).
Added a simple connection tool to show you how I was trying to figure out how to connect to the redis instance (I assumed, wrongly that I should just connect to localhost:6379, turns out there is some magical DNS stuff that you just "have to know").
Lastly, I wrote a simple build script that runs the docker build command followed up with a forced reinstall of the pod, it takes almost a minute to run this, yay.